Add timeout to read operation
Occasionally we see the instumentation
process crash but not close the pipe,
which causes this program to block
forever waiting for input.
This change wraps the read in a timeout
which is treated the same as when the pipe
is broken - the report is still generated
but tests in progress will be marked as failed.